Separate execution and incremental execution tests
