+++ packages/supermassive/src/__tests__/execute.test.ts
@@ -1,25 +1,19 @@
 import {
   parse,
-  experimentalExecuteIncrementally as graphQLExecute,
+  execute as graphQLExecute,
   subscribe as graphQLSubscribe,
-  ExecutionArgs as GraphQLExecutionArgs,
-  Kind,
-  ExecutionResult as GraphQLExecutionResult,
-  ExperimentalIncrementalExecutionResults as GraphQLExperimentalExecutionResult,
   GraphQLSchema,
-  DocumentNode,
-  OperationDefinitionNode,
-  OperationTypeNode,
 } from "graphql";
-import { executeWithoutSchema, executeWithSchema } from "..";
-import { makeSchema, typeDefs } from "../benchmarks/swapi-schema";
+import { makeSchema } from "../benchmarks/swapi-schema";
 import models from "../benchmarks/swapi-schema/models";
-import resolvers from "../benchmarks/swapi-schema/resolvers";
-import { ExecutionResult, UserResolvers } from "../types";
-import { forAwaitEach, isAsyncIterable } from "iterall";
-import { ObjMap } from "../jsutils/ObjMap";
-import { PromiseOrValue } from "graphql/jsutils/PromiseOrValue";
-import { extractMinimalViableSchemaForRequestDocument } from "../supermassive-ast/addMinimalViableSchemaToRequestDocument";
+import { createExecutionUtils } from "../__testUtils__/execute";
+
+const {
+  compareResultsForExecuteWithSchema,
+  compareResultForExecuteWithoutSchemaWithMVSAnnotation,
+  drainExecution,
+  graphqlExecuteOrSubscribe,
+} = createExecutionUtils(graphQLExecute, graphQLSubscribe);
 
 interface TestCase {
   name: string;
@@ -132,10 +126,10 @@ const testCases: Array<TestCase> = [
       name
       gender
       birth_year
-      # This is async iterator
-      starships {
-        name
-      }
+      # TODO: separate field / arg for starships as async iterator (today it breaks with graphql-js >= 16)
+      # starships {
+      #  name
+      # }
     }  
     `,
   },
@@ -367,385 +361,6 @@ query Person($id: Int!) {
       throwError: true,
     },
   },
-  // defer
-  {
-    name: "@defer on inline fragment",
-    document: `
-    {
-      person(id: 1) {
-        name
-        ... on Person @defer {
-          gender
-          birth_year
-        }
-      }
-    }
-    `,
-  },
-  {
-    name: "@defer on fragment",
-    document: `
-    {
-      person(id: 1) {
-        name
-        ...DeferredPerson @defer
-      }
-    }
-
-    fragment DeferredPerson on Person {
-      gender
-      birth_year
-    }
-    `,
-  },
-  {
-    name: "@defer multiple fragments",
-    document: `
-    {
-      person(id: 1) {
-        name
-        ... on Person @defer {
-          gender
-        }
-        ...DeferredPerson @defer
-      }
-    }
-
-    fragment DeferredPerson on Person {
-      birth_year
-    }
-    `,
-  },
-  {
-    name: "@defer overlapping fragment",
-    document: `
-    {
-      person(id: 1) {
-        name
-        ... on Person @defer {
-          gender
-        }
-        ...DeferredPerson @defer
-      }
-    }
-
-    fragment DeferredPerson on Person {
-      gender
-      birth_year
-    }
-    `,
-  },
-
-  {
-    name: "@defer nested",
-    document: `
-    {
-      person(id: 1) {
-        name
-        ...DeferredPerson @defer
-      }
-    }
-
-    fragment DeferredPerson on Person {
-      birth_year
-      ... on Person @defer {
-        films {
-          title
-        }
-      }
-    }
-    `,
-  },
-
-  {
-    name: "@defer label simple",
-    document: `
-   {
-      person(id: 1) {
-        name
-        ... on Person @defer(label: "INLINE") {
-          gender
-        }
-        ...DeferredPerson @defer(label: null)
-      }
-    }
-
-    fragment DeferredPerson on Person {
-      birth_year
-    }
-    `,
-  },
-
-  {
-    name: "@defer on list",
-    document: `
-    {
-      person(id: 1) {
-        name
-        films {
-          ... on Film @defer {
-            title
-          }
-        }
-      }
-    }
-    `,
-  },
-
-  {
-    name: "@defer on query with union",
-    document: `
-    {
-      search(search: "Lu") {
-        __typename
-        ...on Person @defer {
-          name
-        }
-        ...on Planet {
-          name
-        }
-        ... on Transport {
-          name
-        }
-        ... on Species {
-          name
-        }
-        ... on Vehicle{
-          name
-        }
-      }
-    }
-    `,
-  },
-
-  // TODO: Does not error in graphql-js even though it should in spec
-  // {
-  //   name: "@defer label as  variable error",
-  //   document: `
-  //   query ($label: String) {
-  //     person(id: 1) {
-  //       name
-  //       ... on Person @defer(label: $label) {
-  //         gender
-  //       }
-  //       ...DeferredPerson @defer
-  //     }
-  //   }
-
-  //   fragment DeferredPerson on Person {
-  //     birth_year
-  //   }
-  //   `,
-  //   variables: {
-  //     inlineLabel: "INLINE",
-  //   },
-  // },
-
-  /// TODO: Does not error in graphql-js even though it should by spec
-  // {
-  //   name: "@defer label duplicate labels",
-  //   document: `
-  //   query  {
-  //     person(id: 1) {
-  //       name
-  //       ... on Person @defer(label: "SAME") {
-  //         gender
-  //       }
-  //       ...DeferredPerson @defer(label: "SAME")
-  //     }
-  //   }
-
-  //   fragment DeferredPerson on Person {
-  //     birth_year
-  //   }
-  //   `,
-  // },
-
-  {
-    name: "@defer if",
-    document: `
-   {
-      person(id: 1) {
-        name
-        ... on Person @defer(if: true) {
-          gender
-        }
-        ...DeferredPerson @defer(if: false)
-      }
-    }
-
-    fragment DeferredPerson on Person {
-      birth_year
-    }
-    `,
-  },
-
-  {
-    name: "@defer crossselection",
-    document: `
-   {
-      person(id: 1) {
-        name
-        gender
-        ... on Person @defer {
-          gender
-        }
-      }
-    }
-    `,
-  },
-
-  {
-    name: "@defer skip/include",
-    document: `
-   {
-      person(id: 1) {
-        name
-        ... on Person @defer {
-          gender @skip(if: true)
-        }
-        ... on Person @defer {
-          birth_year @include(if: false)
-        }
-        ... on Person @defer @skip(if: true) {
-          skin_color
-        }
-        ... on Person @defer @include(if: false) {
-          hair_color
-        }
-      }
-    }
-    `,
-  },
-
-  {
-    name: "@stream basic",
-    document: `
-    {
-      person(id: 1) {
-        name
-        films @stream {
-          title
-        }
-      }
-    }
-    `,
-  },
-
-  {
-    name: "@stream inside @defer",
-    document: `
-    {
-      person(id: 1) {
-        name
-        ...DeferredPerson @defer
-      }
-    }
-
-    fragment DeferredPerson on Person {
-      birth_year
-      ... on Person @defer {
-        films @stream {
-          title
-        }
-      }
-    }`,
-  },
-
-  {
-    name: "@stream if",
-    document: `
-    {
-      person(id: 1) {
-        name
-        films @stream(if: true) {
-          title
-        }
-        starships @stream(if: false) {
-          name
-        }
-      }
-    }
-    `,
-  },
-
-  {
-    name: "@stream multiple",
-    document: `
-      {
-        person(id: 1) {
-          name
-          films @stream {
-            title
-          }
-          starships @stream {
-            name
-          }
-        }
-      }
-      `,
-  },
-
-  {
-    name: "@stream nested",
-    document: `
-      {
-        person(id: 1) {
-          name
-          films @stream {
-            title
-            planets @stream
-          }
-        }
-      }
-      `,
-  },
-
-  {
-    name: "@stream label",
-    document: `
-      {
-        person(id: 1) {
-          name
-          films @stream(label: "FILMS") {
-            title
-          }
-          starships @stream(label: "STARSHIPS") {
-            name
-          }
-        }
-      }
-      `,
-  },
-  // TODO: not throwing in graphql-js but should by spec - duplicate labels, labels as vars
-
-  {
-    name: "@stream initialCount",
-    document: `
-    {
-      person(id: 1) {
-        name
-        films @stream(label: "FILMS", initialCount: 2) {
-          title
-        }
-      }
-    }
-    `,
-  },
-
-  {
-    name: "@stream/defer errors",
-    document: `
-      {
-        person(id: 1) {
-          name
-          ... on Person @defer {
-            bubblingError
-          }
-          bubblingListError @stream
-        }
-      }`,
-  },
 ];
 
 describe("graphql-js snapshot check to ensure test stability", () => {
@@ -798,131 +413,3 @@ describe("executeWithoutSchema - minimal viable schema annotation", () => {
     );
   });
 });
-
-async function compareResultsForExecuteWithSchema(
-  schema: GraphQLSchema,
-  query: string,
-  variables?: Record<string, unknown>,
-) {
-  expect.assertions(1);
-  const document = parse(query);
-  const result = await drainExecution(
-    await executeWithSchema({
-      typeDefs,
-      resolvers: resolvers as UserResolvers<any, any>,
-      document,
-      contextValue: {
-        models,
-      },
-      variableValues: variables,
-    }),
-  );
-  const validResult = await drainExecution(
-    await graphqlExecuteOrSubscribe({
-      document,
-      contextValue: {
-        models,
-      },
-      schema,
-      variableValues: variables,
-    }),
-  );
-  expect(result).toEqual(validResult);
-}
-
-async function compareResultForExecuteWithoutSchemaWithMVSAnnotation(
-  schema: GraphQLSchema,
-  query: string,
-  variables: Record<string, unknown> = {},
-) {
-  expect.assertions(1);
-  const document = parse(query);
-  const result = await drainExecution(
-    await executeWithoutSchema({
-      document,
-      contextValue: {
-        models,
-      },
-      resolvers: resolvers as UserResolvers,
-      schemaFragment: extractMinimalViableSchemaForRequestDocument(
-        schema,
-        document,
-      ),
-      variableValues: variables,
-    }),
-  );
-  const validResult = await drainExecution(
-    await graphqlExecuteOrSubscribe({
-      document,
-      contextValue: {
-        models,
-      },
-      schema,
-      variableValues: variables,
-    }),
-  );
-  expect(result).toEqual(validResult);
-}
-
-function graphqlExecuteOrSubscribe(
-  args: GraphQLExecutionArgs,
-): PromiseOrValue<GraphQLResult> {
-  const operationName = args.operationName;
-  let operation: OperationDefinitionNode | undefined;
-  for (const definition of (args.document as unknown as DocumentNode)
-    .definitions) {
-    switch (definition.kind) {
-      case Kind.OPERATION_DEFINITION:
-        if (operationName == null) {
-          if (operation !== undefined) {
-            throw new Error("Bad operation in test");
-          }
-          operation = definition;
-        } else if (definition.name?.value === operationName) {
-          operation = definition;
-        }
-        break;
-    }
-  }
-  if (!operation) {
-    throw new Error("Bad operation in test");
-  }
-
-  if (operation.operation === OperationTypeNode.SUBSCRIPTION) {
-    return graphQLSubscribe(args);
-  } else {
-    return graphQLExecute(args);
-  }
-}
-
-type GraphQLResult<TData = ObjMap<unknown>, TExtensions = ObjMap<unknown>> =
-  | GraphQLExecutionResult<TData, TExtensions>
-  | AsyncGenerator<GraphQLExecutionResult<TData, TExtensions>, void, void>
-  | GraphQLExperimentalExecutionResult<TData, TExtensions>;
-
-async function drainExecution(
-  result: ExecutionResult | GraphQLResult,
-): Promise<unknown> {
-  let processedResult;
-  if (isAsyncIterable(result)) {
-    processedResult = await drainAsyncGeneratorToArray(result);
-  } else if ("subsequentResults" in result) {
-    processedResult = {
-      ...result,
-      subsequentResults: await drainAsyncGeneratorToArray(
-        result.subsequentResults,
-      ),
-    };
-  } else {
-    processedResult = result;
-  }
-  return processedResult;
-}
-
-async function drainAsyncGeneratorToArray<T>(
-  collection: AsyncGenerator<T, void, void>,
-): Promise<T[]> {
-  const result: T[] = [];
-  await forAwaitEach(collection, (item) => result.push(item));
-  return result;
-}
